refactor: reset 함수 분리 및 selector 지역상태로 변경
